Let the pure Rust impl build unshimmed
